requirements.txt
aiohttp==3.11.14
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.3.0
click==8.1.8
fastapi==0.115.8
fastapi-mail==1.4.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
passlib==1.7.4
pendulum==3.0.0
phonenumberslite==9.0.3
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg2-binary==2.9.10
pyfiglet==1.0.2
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
setuptools==80.8.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlalchemy-seeder==0.3.0
SQLAlchemy-Utils==0.41.2
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
email_validator==2.2.0
PyJWT==2.10.1
python-multipart==0.0.20
